pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scimodhub"
description = "Sci-ModoM UCSC track hub generator"
readme = "README.md"
# license = { file = "LICENSE" }
authors = [{ name = "Etienne Boileau" }]
maintainers = [
  { name = "Etienne Boileau", email = "boileau@uni-heidelberg.de" }
]
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
"pandas",
"pyyaml",
"requests",
"pydantic[email]",
]

# [project.urls]
# Github = "https://github.com/dieterich-lab/scimodom"
# Issues = "https://github.com/dieterich-lab/scimodom/issues"

[project.optional-dependencies]
tests = ["pytest", "pytest-mock", "pytest-datafiles"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "scimodhub.__version__" }

[project.scripts]
scimodhub = "scimodhub.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"